Move `ImmuTableCell` to new Swift package `ImmuTable`
