+++ WordPress/Classes/Utility/ImmuTable/WPImmuTableRows.swift
@@ -1,6 +1,7 @@
 import Foundation
-import WordPressShared
 import Gridicons
+import ImmuTable
+import WordPressShared
 
 struct NavigationItemRow: ImmuTableRow {
     static let cell = ImmuTableCell.class(WPTableViewCellValue1.self)
