+++ WordPress/Classes/ViewRelated/Blog/Sharing/SharingAccountViewController.swift
@@ -1,5 +1,6 @@
-import UIKit
+import ImmuTable
 import Gridicons
+import UIKit
 import WordPressShared
 
 /// Displays a list of available keyring connection accounts that can be used to
