+++ WordPress/Classes/ViewRelated/Blog/Site Settings/LanguageSelectorViewController.swift
@@ -1,3 +1,4 @@
+import ImmuTable
 import UIKit
 import WordPressShared
