+++ WordPress/Classes/ViewRelated/Pages/Controllers/ParentPageSettingsViewController.swift
@@ -1,3 +1,4 @@
+import ImmuTable
 import UIKit
 
 private struct Row: ImmuTableRow {
