+++ WordPress/Classes/ViewRelated/Plans/ViewModels/PlanListRow.swift
@@ -1,3 +1,4 @@
+import ImmuTable
 import UIKit
 import WordPressShared
