+++ WordPress/Classes/ViewRelated/Plugins/Views/CollectionViewContainerRow.swift
@@ -1,3 +1,5 @@
+import ImmuTable
+
 /// A `ImmuTableRow` that contains a `UICollectionView`.
 /// Currently the `CollectionViewContainerCell` is fairly specific to needs of Plugin Directory,
 /// but it should be possible to make it more generic in the future if there's a need to — it was deliberately
