+++ WordPress/Classes/ViewRelated/Stats/Referrer Details/ReferrerDetailsRow.swift
@@ -1,4 +1,5 @@
 import Foundation
+import ImmuTable
 
 struct ReferrerDetailsRow: ImmuTableRow {
     private typealias CellType = ReferrerDetailsCell
