+++ WordPress/Classes/ViewRelated/Stats/Shared Views/GhostViews/StatsGhostTableViewRows.swift
@@ -1,3 +1,4 @@
+import ImmuTable
 import WordPressUI
 
 protocol StatsRowGhostable: StatsHashableImmuTableRow {
