+++ WordPress/Classes/ViewRelated/Support/SupportTableViewController.swift
@@ -1,3 +1,4 @@
+import ImmuTable
 import UIKit
 import WordPressAuthenticator
