+++ WordPress/Classes/ViewRelated/Tools/Time Zone/Views/TimeZoneRow.swift
@@ -1,3 +1,4 @@
+import ImmuTable
 import UIKit
 
 struct TimeZoneRow: ImmuTableRow {
